+++ test/integration/test_cases/ww_noFade/wNoFade_cases/2_des_pol-without/area_V1/04a_d-w_are_V1.mjs
@@ -38,7 +38,7 @@ const testSteps = [
         }
       },
       {
-        duration: 0.5
+        duration: 1
       }
     ),